Bug Fix: glew was not initialzied.
Fixed a bug where the calls to glGenBuffer threw memory access
violation.

Basically, glew was not initialized and as a result, the opengl
extension was not using the newer opengl interfaces. Most likely
it ws falling back to using the old windows opengl implementations.

"OnWindows, you might use opengl32.lib, which is the import library file for opengl32.dll which comes with every windows version since Windows 95.

However, this DLL does not contain the OpenGL implementation you are typically using, but it contains Microsoft's OpenGL 1.1 GDI software rasterizer. The actual OpenGL implementation on windows is provided by an Installable Client Driver (ICD) which comes with your graphics driver. For OpenGL 1.0 and 1.1 functions, opengl32.dll will act as a trampoline and will forward the calls to the actual ICD DLL.

If you want to call any OpenGL function beyond OpenGL 1.1 (and that one is from 1997), you have to use the OpenGL extension mechanism in every case, as opengl32.dll does not provide these entry points at all, and the compiler/linker will of course not find them." .. Collected from https://stackoverflow.com/questions/55267854/why-use-glad-library-for-opengl-initialization